Assignment_r/requirements.txt
# AI-Powered Document Processing - Requirements
# Pinned versions for stability and reproducibility

# PDF Processing
pdfplumber==0.10.3

# LLM Integration
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
# Alternative: anthropic==0.8.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Web Interface
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0

# Optional: For better JSON handling
pydantic==2.5.3

# Optional: For progress bars
tqdm==4.66.1

# Backend API
fastapi==0.108.0
uvicorn==0.25.0
PyPDF2==3.0.1

# Note: Fuzzy matching uses Python's built-in difflib.SequenceMatcher
# No additional dependency required

# MongoDB Atlas Learning System
pymongo>=4.6.0
motor>=3.3.0  # For async operations
dnspython>=2.4.0  # For SRV connection strings
apscheduler>=3.10.0  # For scheduled learning tasks
